Improve view error handling
We have been too eager when swallowing exceptions in your view render
method. I still think it's ok to just swallow IOExceptions because they
mean that the socket connection ended and it's not our fault and we have
no one to report that failure to anyway, they're gone.

For all other exceptions however, us just swallowing them would just
show an empty screen in the end without any ability to intercept that
with spring error handling.

This commit replaces that behavior by wrapping the given exception in
the new de.quinscape.spring.jsview.JsViewRenderingException.

(The errors that occur happen (almost) exclusively in the registered
 JsViewProviders)